# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация API для управления продуктами и категориями",
        "version": "1.0.0"
    },
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить список всех категорий",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Список категорий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Создать новую категорию",
                "operationId": "bb02f0320882db37f99574196577e446",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Категория успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка создания категории",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Категория не была создана"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка создания"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить категорию по ID",
                "operationId": "4cdf41756315109f1efc72787db5830a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные категории",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категория не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Категория не найдена"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка получения категории"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Обновить категорию по ID",
                "operationId": "19954b1f8aa4916b5f032dc28ea36a24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Категория успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категория не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Категория не найдена"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка обновления категории"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Непредвиденная ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Непредвиденная ошибка"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Удалить категорию по ID",
                "operationId": "f1b8af6e84bd52b9594d98bde748ebd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Категория успешно удалена"
                    },
                    "404": {
                        "description": "Категория не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Категория не найдена"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка удаления категории"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Непредвиденная ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Непредвиденная ошибка"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить все продукты для категории",
                "operationId": "17808557cdb6916426e1150658a7860f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список продуктов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryRequest": {
                "title": "CategoryRequest",
                "description": "Запрос на создание/обновление категории",
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "description": "Название категории",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание категории",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "title": "Category",
                "description": "Схема для категории",
                "properties": {
                    "id": {
                        "description": "ID категории",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание категории",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Схема для продукта",
                "properties": {
                    "id": {
                        "description": "ID продукта",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название продукта",
                        "type": "string"
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "number",
                        "format": "float"
                    },
                    "category_id": {
                        "description": "ID категории",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
...
}